refactor: 예약내역 관리 DB 패키지 정리 및 인터페이스 분리
