+++ app/src/main/java/woowacourse/movie/database/BookHistories.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.database
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
