+++ app/src/main/java/woowacourse/movie/database/BookingHistoryRepositoryImpl.kt
@@ -1,19 +1,19 @@
-package woowacourse.movie
+package woowacourse.movie.database
 
 import android.content.ContentValues
 import android.database.sqlite.SQLiteDatabase
-import woowacourse.movie.BookingHistoryDBHelper.Companion.KEY_DATE
-import woowacourse.movie.BookingHistoryDBHelper.Companion.KEY_SEATS
-import woowacourse.movie.BookingHistoryDBHelper.Companion.KEY_TICKET_COUNT
-import woowacourse.movie.BookingHistoryDBHelper.Companion.KEY_TIME
-import woowacourse.movie.BookingHistoryDBHelper.Companion.KEY_TITLE
-import woowacourse.movie.BookingHistoryDBHelper.Companion.KEY_TOTAL_PRICE
-import woowacourse.movie.BookingHistoryDBHelper.Companion.TABLE_NAME
+import woowacourse.movie.database.BookingHistoryDBHelper.Companion.KEY_DATE
+import woowacourse.movie.database.BookingHistoryDBHelper.Companion.KEY_SEATS
+import woowacourse.movie.database.BookingHistoryDBHelper.Companion.KEY_TICKET_COUNT
+import woowacourse.movie.database.BookingHistoryDBHelper.Companion.KEY_TIME
+import woowacourse.movie.database.BookingHistoryDBHelper.Companion.KEY_TITLE
+import woowacourse.movie.database.BookingHistoryDBHelper.Companion.KEY_TOTAL_PRICE
+import woowacourse.movie.database.BookingHistoryDBHelper.Companion.TABLE_NAME
 import woowacourse.movie.movie.MovieBookingInfo
 import woowacourse.movie.movie.MovieBookingSeatInfo
 
-class BookingHistoryRepository(private val database: SQLiteDatabase) {
-    fun getAll(): List<MovieBookingSeatInfo> {
+class BookingHistoryRepositoryImpl(private val database: SQLiteDatabase) : BookingHistoryRepository {
+    override fun getAll(): List<MovieBookingSeatInfo> {
         val bookingInfoGroup = mutableListOf<MovieBookingSeatInfo>()
         database.rawQuery("SELECT * FROM $TABLE_NAME", null).use {
             while (it.moveToNext()) {
@@ -35,7 +35,7 @@ class BookingHistoryRepository(private val database: SQLiteDatabase) {
         return bookingInfoGroup
     }
 
-    fun insert(movieBookingSeatInfo: MovieBookingSeatInfo) {
+    override fun insert(movieBookingSeatInfo: MovieBookingSeatInfo) {
         val record = ContentValues().apply {
             put(KEY_TITLE, movieBookingSeatInfo.movieBookingInfo.title)
             put(KEY_TIME, movieBookingSeatInfo.movieBookingInfo.time)
@@ -47,14 +47,14 @@ class BookingHistoryRepository(private val database: SQLiteDatabase) {
         database.insert(TABLE_NAME, null, record)
     }
 
-    fun clear() {
+    override fun clear() {
         database.execSQL("DELETE FROM $TABLE_NAME")
     }
 
-    fun isEmpty(): Boolean =
+    override fun isEmpty(): Boolean =
         database.rawQuery("SELECT * FROM $TABLE_NAME", null).use { it.count == 0 }
 
-    fun close() {
+    override fun close() {
         database.close()
     }
 }
